Fix dependency
This is a dependency that was in peerDependencies. That mechanism was
originally chosen as a means of allowing the deployment projects to update the
SDK, but that isn't necessary. Those projects will now simply rely on a
transitive dependency from the integration project. In a case where the
deployment project wants to go up a version, it may still make that happend
without peerDependencies.